# studymanager/src/main/resources/openapi/StudyManagerAPI.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: StudyManager API
security:
  - OAuth:
      - read
      - write
servers:
  - url: https://localhost:8080/api/v1
    description: Local Test Server
paths:
...
  /components/{componentType}/{componentId}/api/{slug}:
    parameters:
      - $ref: '#/components/parameters/ComponentType'
      - name: componentId
        in: path
        schema:
          type: string
        required: true
      - name: slug
        in: path
        schema:
          type: string
        required: true
    post:
      tags:
        - components
      operationId: accessModuleSpecificEndpoint
      description: access module specific endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Endpoint not found
        '500':
          description: Error
...
  /studies/{studyId}:
    get:
      tags:
        - studies
      description: Get study information
      operationId: getStudy
      parameters:
        - $ref: '#/components/parameters/StudyId'
      responses:
        '200':
          description: Successfully returned study information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Study'
        '404':
          description: Not found
    put:
      tags:
        - studies
      description: Update a study
      operationId: updateStudy
      parameters:
        - $ref: '#/components/parameters/StudyId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Study'
      responses:
        '200':
          description: Successfully updated study
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Study'
        '400':
          description: Could not update study
        '404':
          description: Not found
    delete:
      tags:
        - studies
      description: Delete a study
      operationId: deleteStudy
      parameters:
        - $ref: '#/components/parameters/StudyId'
      responses:
        '200':
          description: Study deleted
        '404':
          description: Not found
...
  /studies/import/study:
    post:
      tags:
        - importExport
      description: Import Study
      operationId: importStudy
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary

      responses:
        '201':
          description: import was successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Study'
        '400':
          description: bad request

# --- START GOALS
  /studies/{studyId}/goals/config:
    get:
      tags:
        - goals
      description: Getter for the goal configuration
      operationId: getGoalConfig
      parameters:
        - $ref: '#/components/parameters/StudyId'
      responses:
        '200':
          description: Study goal configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGoalConfigData'
        '404':
          description: The referenced study does not exist
        '400':
          description: The parsed studyId is invalid
    put:
      tags:
        - goals
      description: Updates the goal configuration for the study
      operationId: setGoalConfig
      parameters:
        - $ref: '#/components/parameters/StudyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGoalConfig'
      responses:
        '200':
          description: Study goal configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyGoalConfigData'
        '404':
          description: The referenced study does not exist
        '400':
          description: The parsed study goal configuration or study id was invalid
        '409':
          description: If an adherence check is removed that was still used by some GoalTemplate
  /studies/{studyId}/goals/config/categories/topic:
    post:
      tags:
        - goals
      description: Adds a new topic for goals
      operationId: createGoalTopic
      parameters:
        - $ref: '#/components/parameters/StudyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalTopic'
      responses:
        '200':
          description: Study goal topic created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalTopic'
        '400':
          description: The parsed Study goal topic was invalid
  /studies/{studyId}/goals/config/categories/topic/{key}:
    parameters:
      - $ref: '#/components/parameters/StudyId'
      - name: key
        in: path
        required: true
        schema:
          type: string
    put:
      tags:
        - goals
      description: Updates a topic for goals
      operationId: updateGoalTopic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalTopic'
      responses:
        '200':
          description: The updates study goal topic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalTopic'
        '400':
          description: The parsed Study goal topic was invalid
    delete:
      tags:
        - goals
      description: deletes the topic goals with the referenced key
      operationId: deleteGoalTopic
      responses:
        '204':
          description: action deleted successfully
        '400':
          description: bad request
        '409':
          description: conflict - deleting the goal topic is not possible while it is used by a goal template
  /studies/{studyId}/goals/templates:
    post:
      tags:
        - goals
      description: Add GoalTemplate
      operationId: addGoalTemplate
      parameters:
        - $ref: '#/components/parameters/StudyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalTemplate'

      responses:
        '200':
          description: GoalTemplate added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalTemplate'

        '400':
          description: Could not add goal template
    get:
      tags:
        - goals
      description: List goal templates
      operationId: listGoalTemplates
      parameters:
        - $ref: '#/components/parameters/StudyId'
      responses:
        '200':
          description: operation successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GoalTemplate'
        '404':
          description: not found
  /studies/{studyId}/goals/templates/{templateId}:
    put:
      tags:
        - goals
      description: update goal template
      operationId: updateGoalTemplate
      parameters:
        - $ref: '#/components/parameters/StudyId'
        - $ref: '#/components/parameters/GoalTemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalTemplate'
      responses:
        '200':
          description: goal template successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalTemplate'

        '400':
          description: could not update goal template
        '404':
          description: not found
    delete:
      tags:
        - goals
      description: Delete goal template
      operationId: deleteGoalTemplate
      parameters:
        - $ref: '#/components/parameters/StudyId'
        - $ref: '#/components/parameters/GoalTemplateId'
      responses:
        '204':
          description: Successfully deleted goal template
        '404':
          description: not found
        '500':
          description: Error


# ---- END GOALS -----
  /studies/{studyId}/export/study:
    parameters:
      - $ref: '#/components/parameters/StudyId'
    get:
      tags:
        - importExport
      description: Export Study
      operationId: exportStudy
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyImportExport'
        '400':
          description: bad request
...
  /auditlog/study/{studyId}:
    parameters:
      - $ref: '#/components/parameters/StudyId'
    get:
      tags:
        - auditLog
      description: Provides metadata of the audit logs for the requested study
      operationId: getAuditLogMetadata
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogMetadata'
        '404':
          description: not found
...
components:
  schemas:
...
    Study:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        title:
          type: string
        purpose:
          type: string
        participantInfo:
          type: string
        consentInfo:
          type: string
        duration:
          $ref: '#/components/schemas/StudyDuration'
        finishText:
          type: string
        status:
          $ref: '#/components/schemas/StudyStatus'
        start:
          type: string
          format: date
          readOnly: true
        end:
          type: string
          format: date
          readOnly: true
        plannedStart:
          type: string
          format: date
        plannedEnd:
          type: string
          format: date
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
        applicationAccess:
          type: array
          description: Application types which the user have access to
          items:
            type: string
        userRoles:
          type: array
          description: the roles the _current user_ has within _this_ study.
          items:
            $ref: '#/components/schemas/StudyRole'
          uniqueItems: true
          readOnly: true
          minItems: 1
        contact:
          $ref: '#/components/schemas/Contact'
    Contact:
      type: object
      properties:
        institute:
          type: string
        person:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
...
    StudyStatus:
      type: string
      enum:
        - draft
        - preview
        - paused-preview
        - active
        - paused
        - closed
      default: draft
    StudyGroup:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        studyGroupId:
          $ref: '#/components/schemas/Id'
        title:
          type: string
        purpose:
          type: string
        duration:
          $ref: '#/components/schemas/StudyDuration'
        numberOfParticipants:
          type: integer
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
    ObservationGroup:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        observationGroupId:
          $ref: '#/components/schemas/Id'
        title:
          type: string
        purpose:
          type: string
        numberOfParticipants:
          type: integer
          readOnly: true
        numberOfObservations:
          type: integer
          readOnly: true
        numberOfInterventions:
          type: integer
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
      required:
        - title
    StudyDuration:
      type: object
      properties:
        value:
          type: integer
        unit:
          type: string
          description: unit of time
          enum:
            - MINUTE
            - HOUR
            - DAY
...
    Observation:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        observationId:
          $ref: '#/components/schemas/Id'
        studyGroupId:
          $ref: '#/components/schemas/IdReference'
        title:
          type: string
        purpose:
          type: string
        participantInfo:
          type: string
        type:
          type: string
        properties:
          type: object
          additionalProperties: true
        schedule:
          oneOf:
            - $ref: './Event.yaml/#/components/schemas/Event'
            - $ref: './RelativeEvent.yaml/#/components/schemas/RelativeEvent'
          discriminator:
            propertyName: type
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
        hidden:
          type: boolean
        noSchedule:
          type: boolean
          default: false
        reminder:
          type: boolean
          default: false
        observationGroupIds:
          type: array
          description: The set of observation group Ids the observation is part of
          items:
            $ref: '#/components/schemas/IdReference'
          uniqueItems: true
...
    Intervention:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        interventionId:
          $ref: '#/components/schemas/Id'
        studyGroupId:
          $ref: '#/components/schemas/IdReference'
        title:
          type: string
        purpose:
          type: string
        schedule:
          oneOf:
            - $ref: './Event.yaml/#/components/schemas/Event'
            - $ref: './RelativeEvent.yaml/#/components/schemas/RelativeEvent'
          discriminator:
            propertyName: type
        trigger:
          $ref: '#/components/schemas/Trigger'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Action'
        observationGroupIds:
          type: array
          description: The set of observation group Ids the intervention is part of
          items:
            $ref: '#/components/schemas/IdReference'
          uniqueItems: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
    Trigger:
      type: object
      readOnly: true
      properties:
        type:
          type: string
        properties:
          type: object
          additionalProperties: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
    Action:
      type: object
      properties:
        actionId:
          $ref: '#/components/schemas/Id'
        type:
          type: string
        properties:
          type: object
          additionalProperties: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
    GoalTemplate:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        templateId:
          $ref: '#/components/schemas/Id'
        studyGroupId:
          $ref: '#/components/schemas/IdReference'
        observationGroupIds:
          type: array
          description: The set of observation group Ids the goal template is part of
          items:
            $ref: '#/components/schemas/IdReference'
          uniqueItems: true
        title:
          type: string
        participantTitle:
          type: string
        participantInfo:
          type: string
        type:
          type: string
        categories:
          type: object
          properties:
            kind:
              type: string
              format: enum
              enum:
                - outcome
                - behavioral
            topics:
              description: references to one or more keys of topics defined in the study goal configuration
              type: array
              items:
                type: string
          required:
            - kind
            - topics
        adherenceChecks:
          type: array
          items:
            $ref: '#/components/schemas/AdherenceCheckScheduleEnum'
        properties:
          type: object
          additionalProperties: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
      required:
        - studyId
        - type
        - title
        - participantTitle
        - participantInfo
        - categories
    StudyGoalConfig:
      type: object
      properties:
        consents:
          type: object
          properties:
            commitment:
              type: string
            achievability:
              type: string
            understandable:
              type: string
          required:
            - commitment
            - achievability
            - understandable
        schedule:
          type: array
          items:
            type: object
            properties:
              key:
                $ref: '#/components/schemas/AdherenceCheckScheduleEnum'
              time:
                type: string
                format: time
                description: Follows ISO 8601 format for time
      required:
        - consents
        - schedule
    StudyGoalConfigData:
      allOf:
        - $ref: "#/components/schemas/StudyGoalConfig"
        - type: object
          properties:
            topics:
              type: array
              items:
                $ref: '#/components/schemas/GoalTopic'
      required:
        - consents
        - schedule
        - topics
    AdherenceCheckScheduleEnum:
      type: string
      enum:
        - morning
        - noon
        - afternoon
        - evening
        - night
    GoalTopic:
      type: object
      properties:
        key:
          type: string
        title:
          type: string
        description:
          type: string
      required:
        - title
        - description
    StudyId:
      type: integer
      format: int64
      readOnly: true
    Id:
      type: integer
      format: int32
      readOnly: true
    IdReference:
      type: integer
      format: int32
    StudyImportExport:
      type: object
      properties:
        study:
          $ref: '#/components/schemas/Study'
        studyGroups:
          type: array
          items:
            $ref: '#/components/schemas/StudyGroup'
        observationGroups:
          type: array
          items:
            $ref: '#/components/schemas/ObservationGroup'
        observations:
          type: array
          items:
            $ref: '#/components/schemas/Observation'
        interventions:
          type: array
          items:
            $ref: '#/components/schemas/Intervention'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantInfo'
        integrations:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationInfo'
    ParticipantInfo:
      type: object
      properties:
        studyGroup:
          type: integer
        observationGroups:
          type: array
          items:
            type: integer
          uniqueItems: true
    IntegrationInfo:
      type: object
      properties:
        name:
          type: string
        observationId:
          $ref: '#/components/schemas/Id'
...
    StudyRole:
      description: |
        We differentiate between the following roles within a study:
        * `STUDY_ADMIN` (_Study Administrator_): **All** rights to create and manage studies, configure components and assign participants, invite and manage operators and viewers and see study data.
        * `STUDY_OPERATOR` (_Study Operator_): Rights to manage studies, configure components and assign participants, **no rights to see data**.
        * `STUDY_VIEWER` (_Study Viewer_): Rights to see study daƒta only.
      type: string
      enum:
        - STUDY_ADMIN
        - STUDY_OPERATOR
        - STUDY_VIEWER
      externalDocs:
        url: https://more-platform.atlassian.net/browse/MORE-213
...
    AuditLogMetadata:
      description: |
        Gives back current length of audit log entries with an integer and export format
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        length:
          type: integer
          format: int64
...
  parameters:
    ComponentType:
      name: componentType
      in: path
      schema:
        type: string
        enum:
          - action
          - trigger
          - observation
          - goal
      required: true
    StudyId:
      name: studyId
      in: path
      schema:
        $ref: '#/components/schemas/StudyId'
      required: true
...
    GoalTemplateId:
      name: templateId
      in: path
      schema:
        type: integer
        format: int32
      required: true
...
  securitySchemes:
    OAuth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://auth.more.redlink.io/authorize #TODO
          scopes:
            read: read studies
            write: write studies